bug fixes (trainUtils, trainFunc -> tensors shape)
