+++ Datasets_utils.py
@@ -1,6 +1,7 @@
 
 from random import sample
 from torch.utils.data import IterableDataset
+import torch
 
 class SequentialDataset(IterableDataset):
     def __init__(self, x, y, batch_size):
@@ -13,8 +14,6 @@ class SequentialDataset(IterableDataset):
         self.batches = []
         for seg_x, seg_y in zip(x,y):
             seg = self.__create_batchs__(seg_x, seg_y)
-            print(len(seg))
-            print(seg[0][0], len(seg[0][1]), len(seg[0][2]))
             self.batches.extend(seg)
 
         self.batches = sample(self.batches, len(self.batches))
